Move all fixtures into conftest.py
