+++ tests/unittests/queries/mongodb/test_pymongo.py
@@ -1,17 +1,4 @@
-# ruff: noqa: F401, F811
 from superduperdb.datalayer.mongodb.query import Collection
-
-from tests.fixtures.collection import (
-    empty,
-    float_tensors_16,
-    float_tensors_32,
-    random_data_factory,
-    random_data,
-    a_watcher,
-    a_model,
-    a_watcher_base,
-    a_model_base,
-)
 
 
 def test_find(random_data):
